core: pass long_output to neb api for statechange data #2650
currently, the neb api does not pass any longoutput, which idomod could
forward to ido2db then. ido* fakes that using the normal output, which
is pretty much misleading.

as this is a change on the neb api, it's scheduled for the next major
release. compatibility is restored adding the attribute at the end,
where old object typecasts will just cut it off (so those neb modules
using the nagios header files will not know about it, but still work).

refs #2650

Signed-off-by: Michael Friedrich <michael.friedrich@univie.ac.at>