idoutils: do not pass host and port if not set in the config, letting mysql|pgsql drivers autoselect the socket #2698
the libdbi drivers behave differently if you do not pass host and port,
instead falling back to automated socket detection. in order to do that
successfully, it is required not to set the host if NULL provided, and
checking if the port is other than the default 0 which is internally
used this being an integer.

once one or both is not set, the libdbi will not pass this to the loaded
driver for either mysql or pglsq, and the inner logic of libmysqlclient
or libpq will do the trick for you.

still, you can set the dbsocket option prior to those changes, and
define the local db socket yourself, not letting the db driver guess
where it was built-in (which is of course normally true, when using
packages).

thanks to Alexander Wirt for reporting and testing on the Debian
packages.

refs #2698