core/classic ui: make hashfunc more efficient using sdbm #2761
currently, hashfunc runs in a loop adding up the characters.
this invokes a call to strlen on every run which is bad at once.
furthermore, the underlaying hash algorithm could cause collision
behaviour for similar keys.

               for (i = 0; i < strlen(name1); i++)
                       result += name1[i];

moving strlen before the loop helps a bit, but the hash algorithm needs
to be changed to be more effective once and for all.

the actual function is hash(i) = hash(i - 1) * 65599 + str[i]; which
leads to the magic constant of 65599 being a prime. berkely db and gawk
make use of that hash function as well.

http://www.cse.yorku.ca/~oz/hash.html

such enhancements will most likely only affect large scale systems with
a lot of similar keys, like one would generate with sni's
monitoring::testconfig module. either way, this should be adapted for
the core on comments/downtimes as well as statusdata, helping out the
classic ui scaling even better on reading the status data on each call.

props to Gunnar Beutner for finding and patching it.

refs #2761