core: try the most common macros ($ARGn$, $USERn$) first (Andreas Ericsson) #2674
the arg and user macros are most commonly used, as well as do not
require a copy of the original buffer, so this can be easily moved into
the first macro grabbing section, returning early once the macro has
been grabbed, increasing performance a bit.

kudos to Andreas Ericsson for the finding and patching.

refs #2674