core: use binary search when looking up macro names, instead of insane strcmp() loops (Andreas Ericsson) #2675
without the patch, it currently happens like this

on each macro grabbing, a strcmp() happens to see which macro name this
is on each macro grabbing, the cleaning options are determined
strcmp()'s are pretty expensive, especially as macros are called that
frequently (like HOSTADDRESS).

the patch adds 2 things

- shortcuts for the most common used macros (HOSTADDRESS and HOSTADDRESS6)
  which can be strcmp()'ed and returned early, without any special
  cleaning options
- a precomputed array, sorted by macro name, on macro initialization,
  which is then gone through with binary comparison on the actual
  strcmp()'s

sorting the array alphabethical allows to compare it by value, as
strcmp() would then output. you'll check wether it's greater or less
than the mid value of the array. depending on the returned value of
strcmp() you'll either return the key (0), or step down into the logical
half, left or right. such a seek is pretty much more efficient than a
hash table, as the macro names are rather long in direct comparison with
a unique hash id being calculated.

since the previous version ran within a loop through all available
macros (MACRO_X_COUNT = 154) calling strcmp() which results in an O(n)
complexity for each xmacro, it can be reduced to O(log(n)) complexity by
using the precomputed sorted array with the binary search lookups with
strcmp()

requires further tests and profiling though.

refs #2675