core: re-enable state based escalation ranges, using cast tricks #2878
since we already broke the neb abi a while back, and livestatus
will probably never use any icinga headers itsself, we just
don't give a f*ck about that mad situation - we'll put the
attributes at the end of the structs, tricking the typcasts
leaving the neb api intact.

docs did not receive the configure flag either for a strange
reason, so we can remove that safely from the source as well.

refs #2878