classic ui: fix setting send_notification or sticky_ack as GET param has no effect on cmd.cgi acks #2926
now these GET params require 0 or 1 to be passed via url call,
previously one had to just set that option. this is due to keeping
them enabled as we know before, in order not to need to change the
urls everywhere.

this makes it basically possible, to disable send_notification and
sticky_ack via GET params.

refs #2926