core: add new callback NEBTYPE_CONTACTNOTIFICATIONMETHOD_EXECUTE to allow override of sending notifications #2041
this callback can be overridden just before the call to my_system_r()
happens, so we call it 'execute' (afterwards, we already got 'end',
but preserve 'executeD' for a future callback. though, i guess that
noone will use that one.

hopefully this adds a little benefit to neb authors, if not, it just
solves a long lasting issue for now.

refs #2041