idoutils: go on with mysql insert-on-duplicate-key replacement, add safety SELECT before INSERTS (libdbd-mysql seems broken)
 i guess the faulty return of "affected rows" is just plain mysql dbi driver specific.
because in doing so on a mysql shell, it works - everything changed, and affected is
also 1. in doing so with postgresql, it also works.

probably the driver handlers affected and matched differently within the driver layer
itsself. on debian wheezy i got

ii  libdbd-mysql 0.8.2-1-4.1+b1

need to recheck on wheezy/sid if the error is consistent.

i don't have time to debug the shitty dbd driver though.

in order to stay safe, and not to lose too much performance (because status UPDATES
happen n-1 of the time, while the initial insert only happens once on initial startup
- since we do not clean any *status tables at all), we'll double check this for sanity.

- run the UPDATE
- check if num_rows_affected turns out to be 0 (for whatever reasons later on)
- if it is really 0, do a SELECT on the unique constraint
- the the select is 0, INSERT a new row

this will move the SELECT (which is another performance killing query) out of the
standard-to-be-executed queries, and adds it as safety check before inserting anything
into the table which would violate the unique constraints.

currently done for

- comments
- commenthistory
- hoststatus
- servicestatus
- programstatus

though there might be more tables and queries affected by that, so we should possibly add
safety calls to each query former known as "insert-on-duplicate-key-update-mysql-foobar".

MySQL, you are teh suxx!

refs #3008