core: fixed when deleting an expiring ack the event to delete the ack doesn't get deleted #3084
refs: #3084

added function "delete_scheduled_event" to events.c. This function is only
used to delete acknowledgements at the moment.

when a acknowledgement get's deleted, then we check if the ack has an expire time.
if though then it tryes to delete the event which is resposible to delete the ack.

Sorry I forgot this in first place.

whatthecommit: I am even stupider than I thought