classic ui: javascript refresh jumps to page top, not restoring previous scrolling position; add url param 'scroll=' #2855 - MF
the javascript refresh introduced in previous versions did one thing
wrong - the client side does not know about the current scrolling
position, and therefore the javascript reload was jumping always to
top. being on the client, you cannot store values on reloads in sessions
or similar, but would have to use the cookie method (which is ugly).
a different approach is to add a new GET param, scroll, which gets set
and interpreted by the javascript refresh methods.

since we cannot modify dom objects when reloading the page, but only
when fully loaded, we also need to register eventhandlers in order to
restore the scroll position when we are allowed to.

since the value getters of the actual scroll position are somewhat
different (hey internet exploder 6), this will require more tests,
but the initial implementation should take care of the 3 different
object providers in the first place.

refs #2855