re-add deleted logos/icons for pnp graphs
these cannot be removed from the logos/ directory, as icon_image expects
them to be there (the nagios-images packages from debian does that as
well).

so in order to stay sane with the existing documentation, as well as
addons counting on that location, i'll re-add those icons.

pls deny all future deletion requests as long i am an icinga core dev.