fix faulty macro cleaning, replacing spaces with pluses where they shouldn't be cleaned (refs #3397)
leftover from the macro speedup rewrite - the 0 initialize was missing,
and under random circumstances the previous value was taken, making it
impossible to determine which macros to clean, and which not (as this is
now computed when initializing the macros, and not when grabbing their
values).

added a testcase as well.