idoutils: move the insert id / sequence stuff into ido2db_query_insert_or_update_notificationdata_add (refs #3408)
that qualifies the call on dbi_conn_sequence_last (or ocilib ones) just
for valid inserts. for the update, we can copy the safety select from
mysql, and put it in place when mysql/pgsql with libdbi actually did an
update - updating the referenced object id we should return in that
case. 0 is completely invalid, and must be avoided at all cost.