classic-ui: fixed backtrack_archives only getting calculated on a daily log rotation basis #3416
refs: #3416

number of backtrack_archives are set to find the initial state of
a host/service when reading historical data. since the log reading
got changed with 1.8 the calculation wasn't correct anymore. Now
it calculates the seconds which have to be substract from
start_time_stamp when reading logs, based on @log_rotation_method
and @ backtrack_archives correctly.