classic ui: Implement #3476, "set_expire_ack_by_default"
refs: #3476

This commit implements proposed feature #3476, with input from
Ricardo (and with thanks), in cgi.cfg which automatically ticks the
"Use Expire Time:" check-box in cmd.cgi.  The default behaviour is
to *NOT* tick the checkbox thereby preserving well-established bevaviour.

The intent of the change is in response to observed administrator behaviour
in large organisations where problems get acknowledged and then never fixed
leading to later system failure.  Expiring acknowledgments will assist in
reminding staff in disparate departments and locations that a problem has
not been fixed in a reasonable amount of time specified by the cgi.cfg
setting "default_expiring_acknowledgement_duration" or a different value
given in the "acknowledge problem" dialog.