classic ui: Implement #3467, "set_ack_notifications"
refs: #3467

This commit implements proposed feature #3467 from "TheCry" to tick
or untick the "Send Notification" checkbox in the Acknowledge Problem
dialog.  The default behaviour is set to leave the checkbox ticked
thereby sending notifications when a problem is acknowledged and
preserving the original behaviour.

Thanks to Ricardo for the suggestion of using "set_{option}" as
syntax.