idoutils: fix long output data causes wrong data in database refs #2342
well, well, well. the original fix for truncating the data for mysql
output was a good idea either way. but the root cause for the data
irritation itsself did happen sooner.

ido2db just treats the data it gets in a sequential way. from within the
debug logs remarked in #2342 it's easy to see that the data is truncated
the wrong way, leaving out essential protocol (idomod->ido2db)
information which then causes wrong data insert/update queries in
ido2db.

idomod has a funny method of handling data truncation: first, it creates
a string containing all the protocol information, and cuts that to
IDOMOD_MAX_BUFLEN if too large. the protocol information in the end gets
lost, and so does the data on the other side.

fix for now is to limit long_output/perf_data in advance.