core: increase performance on host->service relation counter (config reading, flapping)
when a service is checked against its host association we can already
update the relation counter as well as total check interval.

the first is required for the config pre flight check while the second
is used within flapping detection.

original patch by Andreas Ericsson, fixed for usage in Icinga.

refs #3820