core: do not export host/servicegroupmembers as environment macros in large installations
this is only true when environment macros are enabled (not the default),
as well as large installation tweaks must be set too. it's likely that
large installations will expose too heavy member strings for the env,
breaking it all - this is just a workaround for the real problem, which
cannot be fixed by icinga itsself, as we wouldn't know on putenv how big
the environment can be once exexcve calls the envp arguments, altogether
with the command line params and existing environment. for details,
check the issue itsself.

refs #3859