nebmods: fix multiple modules sharing same symbols not receiving callback data
once dlopen() loads modules with the same name, it will make sure to
register only the first symbols, and not overriding the coming ones from
more modules of the same binary.

this is bad, as e.g. multiple idomod's will try to register the same
callback function symbols, and only the first one wins then. this is the
reason why only the first instance_name is getting populated in idoutils
database then too.

the initial handshake of ido2db with idomod only happens for the reason
of calling the nebmodule_init() function symbol (from the first module)
with different arguments, passed by the core. each module should have at
least a different instance_name defined in its idomod.cfg, so the init
function will be called with those, doing a handshake (and insert into
icinga_instances table then) then. but afterwards, no data gets received
from the callbacks not having such a diversity (how should the core then
know anyways, it's a _callback_ function called only once from symbol
space!).

basically, the root issue is an omd patch living for more than 2 years
now, but the old behaviour is now restored with this fix (plus adding a
debug idea stolen from Andreas Ericsson, thanks).
alongside fixing this, the solution is rather simple - make a temp copy
of the binary to some random name, and let dlopen actually load that
into memory. then all the registered function symbols will remain module
based and everything else works with the callbacks and so on.

as this is a behavioural change, it should get a note in Changelog once
merged into release trees.

refs #4199