Issue #2613 -- Enhance Solaris Packaging
   This commit represents changes made to the Solaris packaging
routine to ensure proper install/removal of packages and files
from /etc/icinga (configuration), /var/icinga (state and logs),
and /usr/local/icinga (executables, HTML, and libraries).

   The change to Makefile.in is to ensure that the installation
script can be found as the Solaris package is generated; a Solaris
"preinstall" file is now used to make sure that the "icinga" user
and group is present, and the postinstall now makes sure that
permissions, UIDs and GIDs are icinga:icinga for configuration
and state data.

refs #2613