idoutils: remove timedevents & timedeventqueue support and tables
noone ever used that being a performance killer. and those never having
seen it enabled ask themselves whatever that may be. so we're safe to
just drop it while deprecating it once and for all.

db scripts will kill all existing database items too.

(and we're safe for icinga2 ido based on 1.10 schema as well)

refs #4482