core: add colon to illegal_object_name_chars
on demand macros use the ':' as seperator, which causes #4005 as
not-fixed bug. so by adding this we'll make sure that new installs will
respect that option. also in preparation for icinga2 where the colon is
disallowed at any cost to be used in object names (use display_name
for beautiful names anyways!).

fixes #4227