classic ui: Fix incorrect behaviour of getcgivars() -- #5276
This commit attempts to fix an observed behaviour where the getcgivars()
routine was creating a single list rather than a list of key/value pairs.
One manifestation of this was a QUERY_STRING of "jsonoutput=csvoutput"
(which is technically illegal as both of those are booleans) producing
CSV rather than JSON.  This may also help correct the root-cause of the
off-by-one error in #5251.

refs: #5276