External command to {en.dis}able freshness checks for hosts/services.
This patch enables commands to control freshness checking at both the
host and service level. Sometimes, it is necessary to implement
delayed notifications. In our case, passive notifications are being
sent from vehicles that leave the range of our wireless network coverage
area. When the vehicle leaves, we get a notification that the vehicle is
no longer reachable, but it should return within X hours. We wanted to
implement a delayed failure on only specific vehicles. By using freshness
checking, we wait for a passive check to come in saying the vehicle is gone,
then we enable a freshness check on that specific vehicle service and
change the current state to warning. If the vehicle is gone for more than
X hours, the freshness check will run automatically and change the state
to Critical. Hence, a delayed notification is now available. Furthermore,
this is useful with SNMP traps, where the trap may be sent several times,
but we want the first to be a warning, and the second trap to be a failure.

Refs #5422
Refs #4430

Signed-off-by: Michael Friedrich <Michael.Friedrich@netways.de>