classic ui: Previous fix for issue #5276 broke cmd.cgi -- #5276
The prior fix for issue #5276 broke cmd.cgi and was causing segfaults
and core dumps on Solaris.  The Linux behaviour would likely be
breakage to the "interval" option.  The way that FSF/SGI handle NULL
pointers to reference strings is responsible for the discrepency.

fixes: #5276 (again)
whatthecommit: That was a very log day.  I wonder what'll surface next.