classic-ui: fix status.cgi gets stuck in loop when sorting on "All Unhandled Problems" #5886
with the fix for #4404 it was introduced to sort hosts again before
displaying them. This way "statussort_list" was reused. Due to not
properly resetting "statussort_list" to NULL in "free_sort_list",
"sort_status_data" wasn't working properly as it expects
"statussort_list" to be NULL when called.
this also fixes a bug where hosts were sorted only ascending and
sort option wasn't displayed in "HOST_SERVICE_DETAIL" style.

refs: #5886
refs: #5694
whatthecommit: a few bits tried to escape, but we caught them