This commit instantiates the "set_sticky_acknowledgment" CGI option -- refs #6237
This adds a modern-style option in the the CGIs to either set or
clear the "Sticky Acknowledgment" check-box by default.  There is
a competing "original" option, "sticky_ack", but which was not
fully instantiated in all programs nor very well documented.  The
modern idiom is to have checkbox-options begin with the keyword
"set_".

I strongly recommend deprecating the old option, but this commit
still allows both.  I have not changed "config.c" to include the
old option.  The new option has precedence over the old as evidenced
in cmd.c.

A translation from my English patch to the doco into German is required.
I am, unfortunately, a monoglot, and Google "translate" is seriously
lacking.

This commit tests OK in my Solaris-based environment.  Testing should
be undertaken in a Linux one to ensure compatibility and functionality.