Fix QA problem with "config-option-for-sticky-ack-6237"
This commit fixes a QA-related problem which could make the UI
display confusing things when trying to alter the "Sticky Acknowledgment"
tick-box.

The correct merge sequence for this is to first apply revision 1505fd78
and then immediately apply this one as this one fixes *just* the behaviour
problems.  Just applying this commit will miss infrastructure changes
that were necessary to define the desired behaviours.