Classic UI: add possibility to acknowledge host and all its services with one command #3175
It is now possible to acknowledge a problem host and all its services
with one action. Available in extinfo.cgi and status.cgi with
multi select.

The command handling has been in the core for a while now. Only the GUI part was missing.
This command is disable if backend is not Icinga 1.
Refs: #3175