List host alerts in Most Recent Alerts report. Fix #1599.
In add_archived_event(), service_description is set to NULL for events
of type AE_HOST_ALERT. Later on in display_recent_alerts(), all events
without service_description are skipped. But it's ok for host alerts
to have no related service. We're even prepared to write "N/A" to
the reports service column. So skip events only if they're actually
a service alert.

Seems to be a regression introduced with commit a4f01c32a4cfc04430476741e0cbaa30a80310ee.